Create Self-Cleaning-Room
This room optionally keeps itself relatively free of garbage.  If this.self_cleaning is set, nonplayer objects which are not present in .residents are only allowed to remain for this.delay seconds (default one hour).  This is not perfect; it uses :enterfunc as its trigger instead of forking.  Messages:  @remove_object gets printed to everyone in the room when an object is ejected.  [Note kluge: instead of %n to describe the object being removed, use %[tcurrent_names]; this is so multiple items will be mentioned correctly in one message.].  @self-clean commandline verb toggles self cleaning, :set_self_cleaning permits programmatic shift.